+++ src/components/Transcripts.tsx
@@ -6,18 +6,9 @@ import {
   Spacer,
   Flex,
   Text,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-  ModalFooter,
-  FormControl,
-  FormLabel,
-  ModalOverlay,
-  VStack
 } from '@chakra-ui/react';
-import React, { useEffect, useState } from 'react';
-import trpc, { trpcNext } from "../trpc";
+import React, { useState } from 'react';
+import { trpcNext } from "../trpc";
 import { Transcript } from '../shared/Transcript';
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
@@ -29,9 +20,8 @@ import { ChevronLeftIcon, ChevronRightIcon } from '@chakra-ui/icons';
 import { componentSpacing, sectionSpacing } from 'theme/metrics';
 import replaceUrlParam from 'shared/replaceUrlParam';
 import { sidebarBreakpoint } from 'components/Navbars';
-import ModalWithBackdrop from 'components/ModalWithBackdrop';
 import { MinUser } from 'shared/User';
-import { NameMap } from 'shared/Summary';
+import { SummaryNameMapModal } from './SummaryNameMapModal';
 
 export default function Transcripts({ groupId, groupUsers }: {
   groupId: string
@@ -74,9 +64,9 @@ function LoadedTranscripts({ transcripts: unsorted, groupUsers }: {
     const match = summaries.filter(s => s.summaryKey == key);
     summary = match.length ? match[0] : summaries[0];
   }
-  
+
   const { data: nameMap, refetch: refetchNameMap } = trpcNext.transcripts.getNameMap.useQuery({ transcriptId: transcript.transcriptId });
-  
+
   const [nameMapModal, setNameMapModal] = useState(false);
   const handleNameMapModal = async () => {
     setNameMapModal(!nameMapModal);
@@ -91,14 +81,14 @@ function LoadedTranscripts({ transcripts: unsorted, groupUsers }: {
         >前一次</Button>
         <Spacer />
         <Flex direction={{ base: "column", [sidebarBreakpoint]: "row" }} gap={componentSpacing}>
-          <Select value={transcript.transcriptId} 
+          <Select value={transcript.transcriptId}
             onChange={ev => replaceUrlParam(router, "transcriptId", ev.target.value)}
           >
             {sorted.map((t, idx) => <option key={t.transcriptId} value={t.transcriptId}>
               {`${prettifyDate(t.startedAt)}，${prettifyDuration(t.startedAt, t.endedAt)}${!idx ? "（最近通话）" : ""}`}
             </option>)}
           </Select>
-          {summaries && summary && 
+          {summaries && summary &&
             <Select value={summary.summaryKey}
               onChange={ev => replaceUrlParam(router, "summaryKey", ev.target.value)}
             >
@@ -110,9 +100,9 @@ function LoadedTranscripts({ transcripts: unsorted, groupUsers }: {
         <Spacer />
         <Button variant="ghost" rightIcon={<ChevronRightIcon />}
           isDisabled={transcriptIndex == 0}
-          onClick={() => replaceUrlParam(router, "transcriptId", sorted[transcriptIndex - 1].transcriptId)}        
+          onClick={() => replaceUrlParam(router, "transcriptId", sorted[transcriptIndex - 1].transcriptId)}
         >后一次</Button>
-      </Flex> 
+      </Flex>
       {!summary ? <Loader /> :
         <Card variant="outline" backgroundColor="backgroundLight">
           <CardBody>
@@ -121,7 +111,7 @@ function LoadedTranscripts({ transcripts: unsorted, groupUsers }: {
         </Card>
       }
       {nameMapModal &&
-        <NameMapModal
+        <SummaryNameMapModal
           nameMap={nameMap}
           groupUsers={groupUsers}
           onClose={() => {
@@ -130,104 +120,5 @@ function LoadedTranscripts({ transcripts: unsorted, groupUsers }: {
             refetchNameMap();
           }} />}
     </Flex>
-  );
-}
-
-function NameMapModal({ nameMap, groupUsers, onClose }: {
-  nameMap: NameMap | undefined
-  groupUsers: MinUser[]
-  onClose: () => void,
-}) {
-  const [updatedUserMap, setUpdatedUserMap] = useState<{ [key: string]: NameMap}>({});
-  const [initialUserMap, setInitialUserMap] = useState(updatedUserMap);
-  // If not updated the user.name type, will throw a type mismatch error
-  const hasName = (user: { name: string | null; id: string }): user is { name: string; id: string } => {
-    return user.name !== null;
-  };
-
-  useEffect(() => {
-    let userMap: {[key: string]: NameMap} = {};
-    for (const key in nameMap) {
-      const matchedUser = groupUsers.find(gu => gu.name?.includes(nameMap[key]));
-      if (matchedUser && hasName(matchedUser)) {
-        userMap[key] = matchedUser;
-      } else {
-        userMap[key] = { name: nameMap[key], id: '' };
-      }
-    }
-    setUpdatedUserMap(userMap);
-    setInitialUserMap(userMap);
-  }, [nameMap, groupUsers]);
-
-  const handleSelectChange = (e: React.ChangeEvent<HTMLSelectElement>, handlebar: string) => {
-    const selectedOption = e.target.options[e.target.selectedIndex];
-    const selectedUserName = selectedOption.text;
-  
-    // if placeholder/ undefined value is selected
-    if (!e.target.value) {
-      // update the updatedUserMap
-      setUpdatedUserMap(prevMap => ({
-        ...prevMap,
-        [handlebar]: { name: "", id: "" } // Reset to an empty state
-      }));
-    } else {
-      setUpdatedUserMap(prevMap => ({
-        ...prevMap,
-        [handlebar]: { name: selectedUserName, id: e.target.value }
-      }));
-    }
-  };
-
-  const save = async () => {
-    try {
-      trpc.transcripts.updateNameMap.mutate(
-        Object.entries(updatedUserMap).reduce((acc: { [key: string]: string }, [handlebar, userMap]) => {
-          if (userMap.id) { acc[handlebar] = userMap.id; }
-          return acc;
-        }, {})
-      );
-    } finally {
-      onClose();
-    }
-  };
-
-  return (
-    <ModalWithBackdrop isOpen onClose={onClose}>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>User Selection</ModalHeader>
-        <ModalCloseButton />
-        <FormControl>
-          <ModalBody>
-            <VStack spacing={4}>
-              {Object.entries(updatedUserMap || {}).map(([handlebar, groupUser]) => (
-                <FormControl key={handlebar}>
-                  <FormLabel htmlFor={handlebar}>
-                    <ReactMarkdown>{handlebar + ' &hArr; ' + initialUserMap[handlebar].name}</ReactMarkdown>
-                  </FormLabel>
-                  <Select
-                    id={handlebar}
-                    value={groupUser.id}
-                    placeholder='选择用户'
-                    onChange={(e) => handleSelectChange(e, handlebar)}
-                  >
-                    {groupUsers.map((gu) => (
-                      <option key={gu.id} value={gu.id}>
-                        {gu.name}
-                      </option>
-                    ))}
-                  </Select>
-                </FormControl>
-              ))}
-            </VStack>
-          </ModalBody>
-          <ModalFooter>
-            <Button colorScheme='brand' onClick={save}>
-              保存
-            </Button>
-          </ModalFooter>
-        </FormControl>
-      </ModalContent>
-    </ModalWithBackdrop>
   );
 }
