+++ src/shared/Summary.ts
@@ -1,13 +1,13 @@
 import { z } from 'zod';
 
 export const zSummary = z.object({
-    transcriptId: z.string(),
-    summaryKey: z.string(),
-    summary: z.string(),
-  });
+  transcriptId: z.string(),
+  summaryKey: z.string(),
+  summary: z.string(),
+});
 
 export type Summary = z.TypeOf<typeof zSummary>;
 
-export const zNameMap = z.record(z.string());
+export const zSummaryNameMap = z.record(z.string());
 
-export type NameMap = z.TypeOf<typeof zNameMap>;
+export type SummaryNameMap = z.TypeOf<typeof zSummaryNameMap>;
